refactor(planner): refactor hash join
