+++ src/query/service/src/pipelines/processors/transforms/hash_join/hash_join_state.rs
@@ -51,5 +51,5 @@ pub trait HashJoinState: Send + Sync {
     fn right_join_blocks(&self, blocks: &[DataBlock]) -> Result<Vec<DataBlock>>;
 
     /// Get right semi/anti join results
-    fn right_anti_semi_join_blocks(&self, blocks: &[DataBlock]) -> Result<Vec<DataBlock>>;
+    fn right_semi_join_blocks(&self, blocks: &[DataBlock]) -> Result<Vec<DataBlock>>;
 }
