+++ src/query/service/src/pipelines/processors/transforms/hash_join/probe_join/cross_join.rs
@@ -1,0 +1,73 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_datablocks::DataBlock;
+use common_datavalues::Column;
+use common_datavalues::ConstColumn;
+use common_datavalues::DataSchemaRefExt;
+use common_exception::Result;
+
+use crate::pipelines::processors::transforms::hash_join::ProbeState;
+use crate::pipelines::processors::JoinHashTable;
+
+impl JoinHashTable {
+    pub(crate) fn probe_cross_join(
+        &self,
+        input: &DataBlock,
+        _probe_state: &mut ProbeState,
+    ) -> Result<Vec<DataBlock>> {
+        let build_blocks = self.row_space.datablocks();
+        let num_rows = build_blocks
+            .iter()
+            .fold(0, |acc, block| acc + block.num_rows());
+        if build_blocks.is_empty() || num_rows == 0 {
+            let mut fields = input.schema().fields().to_vec();
+            fields.extend(self.row_space.schema().fields().clone());
+            return Ok(vec![DataBlock::empty_with_schema(
+                DataSchemaRefExt::create(fields.clone()),
+            )]);
+        }
+        let build_block = DataBlock::concat_blocks(&build_blocks)?;
+        let mut results: Vec<DataBlock> = Vec::with_capacity(input.num_rows());
+        for i in 0..input.num_rows() {
+            let probe_block = DataBlock::block_take_by_indices(input, &[i as u32])?;
+            results.push(self.merge_with_constant_block(&build_block, &probe_block)?);
+        }
+        Ok(results)
+    }
+
+    // Merge build block and probe block (1 row block)
+    pub(crate) fn merge_with_constant_block(
+        &self,
+        build_block: &DataBlock,
+        probe_block: &DataBlock,
+    ) -> Result<DataBlock> {
+        let mut replicated_probe_block = DataBlock::empty();
+        for (i, col) in probe_block.columns().iter().enumerate() {
+            let replicated_col = ConstColumn::new(col.clone(), build_block.num_rows()).arc();
+
+            replicated_probe_block = replicated_probe_block
+                .add_column(replicated_col, probe_block.schema().field(i).clone())?;
+        }
+        for (col, field) in build_block
+            .columns()
+            .iter()
+            .zip(build_block.schema().fields().iter())
+        {
+            replicated_probe_block =
+                replicated_probe_block.add_column(col.clone(), field.clone())?;
+        }
+        Ok(replicated_probe_block)
+    }
+}
